ai-task-agent/requirements.txt
openai==1.10.0
PyGithub==2.1.1
pyyaml==6.0.1
python-dotenv==1.0.0
